pyproject.toml
[project]
name = "manimations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "manim>=0.19.2",
    "openai>=1.0.0",
    "google-generativeai>=0.6",
    "pyttsx3>=2.90",
    "fastapi>=0.110",
    "uvicorn>=0.25",
    "pydantic>=2.5",
    "gtts>=2.5",
    "python-multipart>=0.0.9",
    "pdfplumber>=0.11",
    "moviepy>=1.0.3",
    "numpy>=1.26",
    "scipy>=1.11",
    "aiofiles>=23.2.1",
    "python-dotenv>=1.0.1",
    "werkzeug>=1.0.1",
]
